pyproject.toml
[project]
name = "python-project-tools"
version = "0.9.0"
description = "python project tools"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.11.3",
    "pydantic-settings>=2.9.1",
    "orjson>=3.10.16",
]

[project.optional-dependencies]
database = [
    "sqlalchemy>=2.0.40",
    "sqlalchemy-utils>=0.41.2",
    "jsonschema>=4.23.0",
]
levenshtein = [
    "levenshtein>=0.27.1",
]
bags = [
    "appdirs>=1.4.4",
    "bagit>=1.8.1",
]
xml2yaml = [
    "pyyaml>=6.0.2",
    "xmltodict>=0.14.2",
]
